changed asset transfer event subscribe from ethers to sdk
